refactor: update import paths to use alias for components
